feature: Added flatMap, map, mapError and pure operators
