+++ test/src/result_test.dart
@@ -156,6 +156,75 @@ void main() {
       });
     },
   );
+
+  group('Map', () {
+    test('Success', () {
+      final result = Success(4);
+      final result2 = result.map((success) => '=' * success);
+
+      expect(result2.tryGetSuccess(), '====');
+    });
+
+    test('Error', () {
+      final result = Error<String, int>(4);
+      final result2 = result.map((success) => 'change');
+
+      expect(result2.tryGetSuccess(), isNull);
+      expect(result2.tryGetError(), 4);
+    });
+  });
+
+  group('MapError', () {
+    test('Success', () {
+      final result = Success<int, int>(4);
+      final result2 = result.mapError((error) => '=' * error);
+
+      expect(result2.tryGetSuccess(), 4);
+      expect(result2.tryGetError(), isNull);
+    });
+
+    test('Error', () {
+      final result = Error<String, int>(4);
+      final result2 = result.mapError((error) => 'change');
+
+      expect(result2.tryGetSuccess(), isNull);
+      expect(result2.tryGetError(), 'change');
+    });
+  });
+
+  group('flatMap', () {
+    test('Success', () {
+      final result = Success<int, int>(4);
+      final result2 = result.flatMap((success) => Success('=' * success));
+
+      expect(result2.tryGetSuccess(), '====');
+    });
+
+    test('Error', () {
+      final result = Error<String, int>(4);
+      final result2 = result.flatMap(Success.new);
+
+      expect(result2.tryGetSuccess(), isNull);
+      expect(result2.tryGetError(), 4);
+    });
+  });
+
+  group('pure', () {
+    test('Success', () {
+      final result = Success<int, int>(4) //
+          .pure(6)
+          .map((success) => '=' * success);
+
+      expect(result.tryGetSuccess(), '======');
+    });
+
+    test('Error', () {
+      final result = Error<String, int>(4).pure(6);
+
+      expect(result.tryGetSuccess(), isNull);
+      expect(result.tryGetError(), 4);
+    });
+  });
 }
 
 Result<SuccessResult, MyException> getMockedSuccessResult() {
@@ -182,8 +251,7 @@ class MyException implements Exception {
   int get hashCode => message.hashCode;
 
   @override
-  bool operator ==(Object other) =>
-      other is MyException && other.message == message;
+  bool operator ==(Object other) => other is MyException && other.message == message;
 }
 
 @immutable
